requirements.txt
django==1.7
psycopg2
google-auth
google-auth-oauthlib
google-auth-httplib2